[#161496897] Add an init file in test v1 folder to debug script
