+++ .gitignore
@@ -3,7 +3,7 @@ __pycache__/
 *.py[cod]
 *$py.class
 .vscode/
-.pyc
+.pyccd
 
 # C extensions
 *.so
