+++ app/api/v1/views/products.py
