+++ app/tests/v1/test_products.py
@@ -1,6 +1,6 @@
 import unittest
 from flask import json
-from app.apps import create_app
+from app import create_app
 from app.api.v1.models.product import ListDatabase
 
 config_name = "testing"
