+++ requirements.txt
@@ -1,29 +1,1 @@
-aniso8601==4.0.1
-atomicwrites==1.2.1
-attrs==18.2.0
-Click==7.0
-coverage==4.5.1
-Flask==1.0.2
-Flask-RESTful==0.3.6
-Flask-RESTful-extend==0.3.7
-Flask-SQLAlchemy==2.3.2
-funcsigs==1.0.2
-gunicorn==19.9.0
-ItsDangerous==1.0.0
-Jinja2==2.10
-json-encode-manager==0.0.2
-MarkupSafe==1.0
-more-itertools==4.3.0
-pathlib2==2.3.2
 pkg-resources==0.0.0
-pluggy==0.8.0
-psycopg2-binary==2.7.5
-py==1.7.0
-pytest==3.9.2
-pytest-cov==2.6.0
-python-dotenv==0.9.1
-pytz==2018.6
-scandir==1.9.0
-six==1.11.0
-SQLAlchemy==1.2.12
-Werkzeug==0.14.1
