+++ run.py
@@ -1,10 +1,9 @@
-from app.apps import create_app
+from appflask import create_app
 
 # Import create_app in run.py in order to be
 # able to run app.
 
 app = create_app()
-
 
 
 if __name__ == '__main__':
